mediscanAI/VideoAssistant/requirements.txt


jupyter
ffmpeg-python
python-dotenv
shiny
git+https://github.com/jcheng5/shinymedia
pandas



Flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0


requests==2.31.0
gunicorn==21.2.0
numpy>=1.24,<2.3
scipy>=1.10.0,<1.14.0



numpy>=1.24,<2.3
